refactor: relocate version information to a new __version__ module
